prepare playground
